last version 15/03/2024 16:00

+++ database/migrations/2024_02_14_113435_create_produits_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->string('designation');
             $table->string('calibre');
             $table->string('type_quantite');
-            $table->decimal('prix_vente')->nullable();
+            $table->float('prix_vente')->nullable();
             $table->unsignedBigInteger('user_id');
             $table->foreign('user_id')->references('id')->on('users')->onDelete('restrict');
             $table->unsignedBigInteger('categorie_id');
